requirements_streamlit.txt
# Core
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0

# API calls
requests==2.31.0
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# For local API (if deploying API too)
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.19
pydantic==2.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ML
scikit-learn==1.3.0

# Parsers
pytesseract==0.3.10
opencv-python-headless==4.8.0.74
Pillow==10.0.0
PyPDF2==3.0.1

# Currency
forex-python==1.6
